update demo component

+++ src/components/Demo.jsx
@@ -1,49 +1,63 @@
-import { useState, useEffect } from 'react'
+import React, { useState, useEffect } from "react";
 
-import { copy, linkIcon, loader, tick } from '../assets'
-import { useLazyGetSummaryQuery } from '../services/article'
+import { copy, linkIcon, loader, tick } from "../assets";
+import { useLazyGetSummaryQuery } from "../services/article";
 
 const Demo = () => {
   const [article, setArticle] = useState({
-    url: '',
-    summary: ''
-  })
+    url: "",
+    summary: "",
+  });
+  const [allArticles, setAllArticles] = useState([]);
+  const [copied, setCopied] = useState("");
 
-  const [allArticles, setAllArticles] = useState([])
-  const [getSummary, { error, isFetching }] = useLazyGetSummaryQuery()
-  const [copied, setCopied] = useState('')
+  // RTK lazy query
+  const [getSummary, { error, isFetching }] = useLazyGetSummaryQuery();
 
+  // Load data from localStorage on mount
   useEffect(() => {
     const articlesFromLocalStorage = JSON.parse(
-      localStorage.getItem('articles')
-    )
+      localStorage.getItem("articles")
+    );
 
     if (articlesFromLocalStorage) {
-      setAllArticles(articlesFromLocalStorage)
+      setAllArticles(articlesFromLocalStorage);
     }
-  }, [])
+  }, []);
 
   const handleSubmit = async (e) => {
-    e.preventDefault()
-  
-    const { data } = await getSummary({ articleUrl: article.url })
+    e.preventDefault();
 
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
+
+    if (existingArticle) return setArticle(existingArticle);
+
+    const { data } = await getSummary({ articleUrl: article.url });
     if (data?.summary) {
-      const newArticle = { ...article, summary: data.summary }
-      const updatedAllArticles = [newArticle, ...allArticles]
+      const newArticle = { ...article, summary: data.summary };
+      const updatedAllArticles = [newArticle, ...allArticles];
 
-      setArticle(newArticle)
-      setAllArticles(updatedAllArticles)
-
-      localStorage.setItem('articles', JSON.stringify(updatedAllArticles))
+      // update state and local storage
+      setArticle(newArticle);
+      setAllArticles(updatedAllArticles);
+      localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
     }
-  }
+  };
 
+  // copy the url and toggle the icon for user feedback
   const handleCopy = (copyUrl) => {
-    setCopied(copyUrl)
-    navigator.clipboard.writeText(copyUrl)
-    setTimeout(() => setCopied(false, 3000))
-  }
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(false), 3000);
+  };
+
+  const handleKeyDown = (e) => {
+    if (e.keyCode === 13) {
+      handleSubmit(e);
+    }
+  };
 
   return (
     <section className='mt-16 w-full max-w-xl'>
@@ -53,32 +67,32 @@ const Demo = () => {
           className='relative flex justify-center items-center'
           onSubmit={handleSubmit}
         >
-          <img 
+          <img
             src={linkIcon}
-            alt='link_icon'
+            alt='link-icon'
             className='absolute left-0 my-2 ml-3 w-5'
           />
 
-          <input 
+          <input
             type='url'
-            placeholder='Enter a URL'
+            placeholder='Paste the article link'
             value={article.url}
-            onChange={(e) => {setArticle({ ...article, url: e.target.value })}}
+            onChange={(e) => setArticle({ ...article, url: e.target.value })}
+            onKeyDown={handleKeyDown}
             required
-            className='url_input peer'
+            className='url_input peer' // When you need to style an element based on the state of a sibling element, mark the sibling with the peer class, and use peer-* modifiers to style the target element
           />
-
           <button
             type='submit'
-            className='submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700'
+            className='submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700 '
           >
-            ↵
+            <p>↵</p>
           </button>
         </form>
 
-        {/* Browese URL History */}
+        {/* Browse History */}
         <div className='flex flex-col gap-1 max-h-60 overflow-y-auto'>
-          {allArticles.map((item, index) => (
+          {allArticles.reverse().map((item, index) => (
             <div
               key={`link-${index}`}
               onClick={() => setArticle(item)}
@@ -87,7 +101,7 @@ const Demo = () => {
               <div className='copy_btn' onClick={() => handleCopy(item.url)}>
                 <img
                   src={copied === item.url ? tick : copy}
-                  alt='copy_icon'
+                  alt={copied === item.url ? "tick_icon" : "copy_icon"}
                   className='w-[40%] h-[40%] object-contain'
                 />
               </div>
@@ -99,14 +113,14 @@ const Demo = () => {
         </div>
       </div>
 
-      {/* Display Results */}
+      {/* Display Result */}
       <div className='my-10 max-w-full flex justify-center items-center'>
         {isFetching ? (
           <img src={loader} alt='loader' className='w-20 h-20 object-contain' />
         ) : error ? (
           <p className='font-inter font-bold text-black text-center'>
             Well, that wasn't supposed to happen...
-            <br/>
+            <br />
             <span className='font-satoshi font-normal text-gray-700'>
               {error?.data?.error}
             </span>
@@ -127,7 +141,7 @@ const Demo = () => {
         )}
       </div>
     </section>
-  )
-}
+  );
+};
 
-export default Demo
+export default Demo;
